move e2e to the test folder
