+++ test/e2e/client/client.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/cloud-provider-alibaba-cloud/e2e/options"
 	ctrlCfg "k8s.io/cloud-provider-alibaba-cloud/pkg/config"
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/provider/alibaba"
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/provider/alibaba/base"
@@ -12,6 +11,7 @@ import (
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/provider/alibaba/pvtz"
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/provider/alibaba/slb"
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/provider/alibaba/vpc"
+	"k8s.io/cloud-provider-alibaba-cloud/test/e2e/options"
 	"k8s.io/klog/v2"
 	"os"
 	runtime "sigs.k8s.io/controller-runtime/pkg/client"
