+++ test/e2e/testcase/route/route.go
@@ -6,9 +6,9 @@ import (
 	"github.com/onsi/gomega"
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/cloud-provider-alibaba-cloud/e2e/framework"
-	"k8s.io/cloud-provider-alibaba-cloud/e2e/options"
 	ctrlCfg "k8s.io/cloud-provider-alibaba-cloud/pkg/config"
+	"k8s.io/cloud-provider-alibaba-cloud/test/e2e/framework"
+	"k8s.io/cloud-provider-alibaba-cloud/test/e2e/options"
 	"strings"
 	"time"
 )
