+++ test/e2e/testcase/service/loadbalancer.go
@@ -5,10 +5,10 @@ import (
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/gomega"
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/cloud-provider-alibaba-cloud/e2e/framework"
-	"k8s.io/cloud-provider-alibaba-cloud/e2e/options"
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/controller/service"
 	"k8s.io/cloud-provider-alibaba-cloud/pkg/model"
+	"k8s.io/cloud-provider-alibaba-cloud/test/e2e/framework"
+	"k8s.io/cloud-provider-alibaba-cloud/test/e2e/options"
 )
 
 func RunLoadBalancerTestCases(f *framework.Framework) {
