--- test/svc.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    run: metaserver
-  name: metaserver
-  namespace: default
-spec:
-  ports:
-  - nodePort: 30977
-    port: 80
-    protocol: TCP
-    targetPort: 80
-  selector:
-    run: metaserver
-  type: NodePort
