--- test/udp.svc.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  labels:
-    run: nginx
-  name: nginx-health-check-udp-test
-  namespace: default
-spec:
-  externalTrafficPolicy: Cluster
-  ports:
-  - port: 8080
-    protocol: UDP
-    targetPort: 80
-  selector:
-    run: nginx
-  type: LoadBalancer
